refactor: update imports to use shadcn directory in payments pages
